Added MinorMod speed boost with shift key
Also made it so the animation speed of the player changes when they sprint.
In Randall_Main-Copy scene i experimented with it as i had the player object but i was using a different script on it. Once i was satisfied, i moved it onto the Player object in the main scene